Add Dynamic DOM Loading Helpers
I added a few helpers for dynamically populating the DOM to the
javascript utils file.

I've been using these helpers for the common task of adding a tree of
nodes to a specific place in the layout or replacing a specific section
with some other dynamic content.

I also included a version of each helper that adds a loading spinner
while the dynamic content is loaded. I don't usually use these in my
sites since I prefer to diagnose the bottleneck and make the page load
fast enough that it isn't noticeable, but there are some situations
where having a closure like this that can automatically set up a loading
spinner is useful, so I have included these variants in the utils file
for future copy/pasta when I need them.